Adds Contour (#3)
TL;DR
-----

Puts the Contour install under GitOps

Details
-------

Switches to this repository to install Contour in the management
cluster. It was previously installed as a Tanzu package but I wanted
to shift it to a GitOps install. I left the install as the traditional
Ingress installation instead of moving toward Gateway API until I
better understood the implications for the services running in the
cluster.