fix sops key config (#7)
* Corrects SOPS config

TL;DR
-----

Updates the SOPS configuration and re-encrypts API key secret

Details
-------

Fixes a defect where the Cloudflare API key secret wasn't properly
decrypted in the cluster because it was encryped with the wrong
key.

* Completes earlier fix

TL;DR
-----

Updates kustomization for renamed manifest file

Details
-------

Corrects the kustomization for cert-manager. I had corrected a file
name typo but not updated the kustomization to refer to the right
file. This change fixes that.