Updates API key value (#8)
Updates API key value

TL;DR
-----

Corrects the value for the API key in the Cloudflare API key secret

Detals
------

Switches the API Key secret to use `stringData` over `data` and
assures it is set to the correct value. I'm making this change since
the kustomization was reporting issues with the type of the secret
data. Not 100% sure this is the solution, but after aligning the
GPG keys it's the best thing I can think of to get around the issue
I'm having with decryption (especially given the error message of
`error: data values must be of type string...`.