Configures shared services cluster (#9)
TL;DR
-----

Provides the base configuraton and boot start for a shared services cluster

Details
--------

Creates a Tanzu cluster to use as a cluser for shared platform services modeled
on the one in [TKG Lab](/Tanzu-Solutions-Engineering/tkg-lab). I'm planning to
vary some many of the services but it will play the same role.  I used a dry run 
of the cluster I used to install using the procedure from that repository as
the base for the files added.  One change I made from the dry run was to split 
the secrets into their own file to encrypt with SOPS.

The created cluster is bootstrapped with the baseline components from a [new
repositoryt](/shortrib-labs/cluster) that I'm going to use to provide common
cluster infrastructure like cert-manager and Contour. I modeled that cluster
on how I installed those components in this one. Likely I'll refeactor later
so that the management cluster also uses that baseline. I'll aslo start to 
enrich the shared services cluster by adding more infrastructue there.