Trims down to management cluster specific manifests
TL;DR
-----

Removes files that no longer belong in this repository

Details
-------

Reworks this repostitory for its new purpose as the source of
truth for workloads that are only run in management cluster. Files
specific to an invividual cluster instance, or common to all
clusters, are now managed in different repositories.

The repositores that make up the lab going forward are:

* [Kubenetes Lab](/shortib-labs/kubernetes-lab] provides fro a
  management cluster managed by kURL and specifies the configuration
  for all clustersi the lab.
* [Management Cluster](/shortrib-labs/management-cluster) this
  repository that specifies workload unique to a management cluster
* [Shared Services Cluster](/shortrib-labs/shared-services-cluster)
  specifies workloads unique to a shared services cluster based on
  [TKG Lab](/Tanzu-Solution-Engineering/tkg-lab] and
  [FRSCA](/buildsec/frsca).
* [Baseline Cluster](/shortrib-labs/baseline-cluster) describes
  workloads common to all clusters